Rename chat_model.function_schemas to chat_model.function_schemas
