+++ src/magentic/chat_model/function_schema.py
